control ping loss quality with existing checkboxes
Instead of specifying a maximum ping loss ratio for detecting connected spans, use the
same logic that is used to determine if a ping loss rectangle should be displayed.
This removes the ability to say ping loss above an arbitrary ratio is "bad", but makes
it possible to see how spans are affected by each category of ping loss.