consider downlink/uplink throughput before classifying outage
There has been some speculation that pings may be so low in priority
that high enough downlink or uplink traffic may cause them to be lost.
This could mean that all-pings-lost seconds are not outages, but instead
just bandwidth devoted to other traffic. This change introduces controls
for specifying what level of downlink/uplink traffic to consider high
enough to overrule a high level of ping loss.